+++ api/users.ts
@@ -1,5 +1,6 @@
 import { toast } from "react-hot-toast";
 import axios, { AxiosError } from "axios";
+
 import { UserInfo } from "../types";
 
 export const getUser = async (token: string | null) => {
@@ -28,27 +29,13 @@ export const postUser = (userInfo: UserInfo, callback?: () => void) => {
     });
 };
 
-export const login = (
-  email: string,
-  password: string,
-  callback?: () => void,
-) => {
-  axios
-    .post(
-      `${process.env.NEXT_PUBLIC_API_URL}/users/login`,
-      { email, password },
-      { withCredentials: true },
-    )
-    .then((res) => {
-      localStorage.setItem("debate-ducks-token", res.data);
-      if (callback) callback();
-      toast.success("로그인에 성공했습니다.");
-    })
-    .catch((err: AxiosError<{ message: string }>) => {
-      toast.error(
-        `${err.response?.data?.message || "네트워크 에러가 발생했습니다."}`,
-      );
-    });
+export const login = async (userInfo: Omit<UserInfo, "name">) => {
+  const { data } = await axios.post(
+    `${process.env.NEXT_PUBLIC_API_URL}/users/login`,
+    { email: userInfo.email, password: userInfo.password },
+    { withCredentials: true },
+  );
+  return data;
 };
 
 export const patchUserImage = async (
